scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.11.4",
    "pyyaml>=6.0.2",
    "sqlmodel>=0.0.24",
    "typer>=0.15.3",
]

[tool.uv]
dev-dependencies = [
    "maturin>=1.8.6",
    "mypy>=1.15.0",
    "pydantic-settings>=2.9.1",
    "types-pyyaml>=6.0.12.20250402",
]

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"
